+++ Justific.Dominio/Dtos/ItemListaOrganizacaoUsuarioDto.cs
@@ -4,9 +4,9 @@ namespace Justific.Dominio.Dtos
 {
     public class ItemListaOrganizacaoUsuarioDto
     {
-        public int OrganizacaoId { get; set; }
+        public long OrganizacaoId { get; set; }
         public string NomeOrganizacao { get; set; }
-        public int UsuarioId { get; set; }
+        public long UsuarioId { get; set; }
         public string LoginUsuario { get; set; }
     }
 }
